refactor: OrderStatus enum common으로 이동
