+++ backend/src/main/java/easy/gc_coffee_api/entity/common/OrderStatus.java
@@ -1,4 +1,4 @@
-package easy.gc_coffee_api.entity;
+package easy.gc_coffee_api.entity.common;
 
 public enum OrderStatus {
     PENDING, SHIPPED
